Changed setReportData length to 1 byte

+++ src/ImprovedKeyboard.cpp
@@ -116,7 +116,7 @@ void Keyboard_::send_now(void){
 }
 
 #if defined(HID_KEYBOARD_LEDS_ENABLED)
-void Keyboard_::setReportData(const void* data, int len){
+void Keyboard_::setReportData(const void* data, uint8_t len){
     // Save led state
     if(len == 1)
     	leds = *(uint8_t*)data;
