+++ src/ImprovedKeyboard.h
@@ -57,7 +57,7 @@ private:
   HID_KeyboardReport_Data_t _keyReport;
   void sendReport(HID_KeyboardReport_Data_t* keys);
 #if defined(HID_KEYBOARD_LEDS_ENABLED)
-  virtual void setReportData(const void* data, int len);
+  virtual void setReportData(const void* data, uint8_t len);
   uint8_t leds;
 #endif
 public:
